Move Fault data types into a separate file
